+++ CHANGELOG.md
@@ -8,3 +8,4 @@
   - feature: auto clean the invalid keys in `permission.json` file, for clarity.
   - feature: sort the `command path` alphabetically, making it easier to read.
 - [command_cooldown] feature: enhanced `/command-cooldown list` command, now it also list the `unnamed cooldown` in details.
+- [core] feature: improve the `game profile` fetch function.